+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/TLMessage.java
@@ -234,7 +234,7 @@ public class TLMessage extends TLAbsMessage {
     }
 
     @Override
-    public int getConstructorId() {
+    public long getConstructorId() {
         return CONSTRUCTOR_ID;
     }
