+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/TLPrivacyKeyStatusTimestamp.java
@@ -19,7 +19,7 @@ public class TLPrivacyKeyStatusTimestamp extends TLAbsPrivacyKey {
     }
 
     @Override
-    public int getConstructorId() {
+    public long getConstructorId() {
         return CONSTRUCTOR_ID;
     }
 }
