+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/request/TLRequestAccountUpdateDeviceLocked.java
@@ -73,7 +73,7 @@ public class TLRequestAccountUpdateDeviceLocked extends TLMethod<TLBool> {
     }
 
     @Override
-    public int getConstructorId() {
+    public long getConstructorId() {
         return CONSTRUCTOR_ID;
     }
