+++ tl/src/main/java/com/github/badoualy/telegram/tl/exception/InvalidConstructorIdException.java
@@ -2,9 +2,9 @@ package com.github.badoualy.telegram.tl.exception;
 
 public class InvalidConstructorIdException extends DeserializationException {
 
-    public InvalidConstructorIdException(int foundId, int expectedId) {
-        super("Invalid constructor id. Found " + Integer.toHexString(foundId)
-                      + ", expected: " + Integer.toHexString(expectedId));
+    public InvalidConstructorIdException(long foundId, long expectedId) {
+        super("Invalid constructor id. Found " + Long.toHexString(foundId)
+                      + ", expected: " + Long.toHexString(expectedId));
     }
 
     public InvalidConstructorIdException(String message) {
