+++ tl/src/main/java/com/github/badoualy/telegram/tl/exception/UnsupportedConstructorException.java
@@ -2,7 +2,7 @@ package com.github.badoualy.telegram.tl.exception;
 
 public class UnsupportedConstructorException extends DeserializationException {
 
-    public UnsupportedConstructorException(int constructorId) {
-        super("Unsupported constructor: #" + Integer.toHexString(constructorId));
+    public UnsupportedConstructorException(long constructorId) {
+        super("Unsupported constructor: #" + Long.toHexString(constructorId));
     }
 }
